fix(/spring_1_0):The final node has been corrected
Change details:

The changes were made to correct the initial and final nodes in the
following files: spring_1_0_stages.py and
retracting_2_stage.py.

The initial node has been set to -1, while the final node is now -2.
These nodes are specifically identified by the ESP as the initial and
final nodes.