log: Refactor logging into its own class
The existing logging logic manually rotates the logs on application start.
The old logs are appended and the new logs go into a new file.
The maximum amount of logs was 99999 files.
This comes with multiple problems: Not only can we not use the auto formating
capabilities of pythons logging modules but we are also unable to control
the maximum amount of discspace the logs can occupy.

For rotating logs python itself offers the RotatingFileHandler class.
We can as for logs to automatically rotate after 200MiB and a maximum amount of logs
which we hardcode to 10 for now. This means that we have a hard upper bar of how much
discspace the logs can occupy. To reduce this even further we are additionally
compressing the logs with gzip. As we are printing logs with a lot of pretetition we
can be reasonably sure that the compressed logs will always be smaller than the
original files and therefore will never be more than 2GiB in total.
On top of that the newly introduced MeticulousLogger class takes the modules name
as a parameter so that we get a good overview which file / modules has been the source
of a speicific log.
Due to the nature of using log handlers we can log to disc and stdout at the same time
and are not longer required to differentiate between print() and "log()" or "add_to_buffer()"
respectively.
The logging is currently happening blocking with the file writes being async on OS level.
Blackbox testing did not show significant performance differences, however we might need to revisit
this behaviour in the future again.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>