esp: Move connection handeling into its own class
So far the ESP32's connection-handling lived inside the back.py main file.
This comes with a multitude of downsides, including the inability to unit
test certain components but also the dependencies on the specific GPIO
pins that the current fika board uses. It is therefore impossible to connect
a USB<->UART connector to the fika board to just talk to the ESP32 from another
machine for development efforts.
Move the connection logic out of the main file and into its own class for
reusability and encapsulation.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>