esp: Move flashing logic back into esptool without closing the port
Flashing the ESP32s is always handled via UART. When the ESP32
boots the ROM checks if GPIO-0 is pulled low and if so it starts the
serial programmer mode.
This mode is in almost all cases implemented by esptool.py or its new
sibling esp-serial-flasher (writen in C).
We can therefore be reasonable sure that we will always have to use
one of those tools to programm the ESP32.
esptool.py is currently called after a multitude of components have
acted on the signal from the back.py code.

Given its python nature we can simply call esptool.py from within
the backend itself, instead of informing a listener which is then
calling a programm which in return is calling the wrapper CLI interface
to that exact software anyways.
As an added bonus we do not need to reconnect to the ESP32. The UART
connection is just passed to esptool instead.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>