esp: Add USBSerialConnection for standalone ESP32 board
With the connection to the Fika Board move to its own class
we can now create a new connection type for ESP32 devboards and
the fika board when it is connected to a host computer rather than
the SOMs UART.
The USBSerialConnection is using esptool.py to reset the ESP32 over
calling individual GPIO pins.
We descided to break out the logic here and not extend the Fika
implementations pyserial port with the required logic to call the
custom pins to be later able to add a dummy connection which will stream
prerecorded data.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>