wifi: Add WifiManager class to connect to a wifi and get system status
To be able to use the machine in production we need to be able to
provision wifi. Instead of using the NetworkManager dbus API we are
using a wrapper around the nmcli tool.
This allows us to keep stability across different python versions
as the respective python dbus projects are currently not very
stable.

So far the WifiManager supports:
- Connecting to WPA-2/3 with PSK
- Scan for available networks
- Parsing the current network configuration into a dataclass

Signed-off-by: Mimoja <mimoja@meticuloushome.com>