wifi: Add BLE GATT server for wifi provisioning
To provision the machines wifi we need a way to get the configuration
data into the machine. As wifi is the system to be configured and
cables are not an option we are utilizing BLE via the Bless modules
which wraps around BlueZs dbus API.

Utilizing the Improv (https://www.improv-wifi.com) protocol
it allows to provision the SOMs wifi connection for WPA2/3 but not
for WPA-Enterprise.
To support this we will later either need to extend the Improv
protocol or add a http configuration interface to be used with the
upcoming standalone WIFI-AP mode.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>