main: Move connection logic into the main try/catch block
We are already running the main() function within a try-catch block.
However multiple components we initialize early are also able to
misbehave.
To no longer crash without logs we are moving all initialization logic
into the try-catch and close the programm gracefully afterwards.
This also allows for external tools (like systemd) to determin the
exit status and restart the backend.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>