main: Remove pynput dependency as it is no longer functional
The pynput modules only functions on linux when used against an
xorg x-server. It does not work with wayland where the only easy
way is to use the kernels uInput subsystem to emulate a keyboard.
We therefore replace the pynput modules with calls to ydotool as
a bandaid to ensure we can use the dial app until we have a proper
socket.io connection.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>