log: Readd sensor logging during machine usage
When we moved the logging into its own class we removed the mayority of
duplicate print() statements in favour of info and debug loggings.
While doing so we accidentally removed the logic to only print when the
machine is busy as initial experiements had shown that we would always
print.
Readd the old logic as it is currently is use!

Fixes: fca4a4f95a1  log: Refactor logging into its own class

Signed-off-by: Mimoja <mimoja@meticuloushome.com>