sensors: Introduce data classes for sensor data parsing
Up until this point the parsing of the serial protocol happened
in the main file back.py, leading to a crowded and difficult to read
file.
As sensor values are owned by the esps firmware there is no reason for
them to be writable in the backend.
We therefore move them into read-only dataclasses.
The sensor data itself is never handled in the backend but instead only
passed to the respective frontends via socket.io.
The mapping for this was created when sending also in back.py.
We move it to the respective data_classes for now so that we can later
manipulate the mappings at one central point if we so desire.

Dataclass fields do allow for additional metadata which in the future we can
and should use for the sio mapping

Signed-off-by: Mimoja <mimoja@meticuloushome.com>