sensors: Send button events via socketio
Button events are currently send via keyboard inputs.
This is not only slow but also unefficient as we have to traverse
uInput, libinput and other componenets on the way to the dial app.
On top of that we are unable to send metadata on button events
as the a keypress is only ever up or down.
To get rid of the keyboard dependency we emit the button events
additionally to the keyboard events so that the frontend team
can develop aginst those.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>