chore: Remove watcher communication
The watcher is not serving a purpose in our final design. It will be able
to update the debian packages in case we break the backend. But restarting
the backend is not beneficial - we are not tracking our threads properly
and can not make use of any monitoring anyways.
A different system needs to be build.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>