system: Create Dockerfile and docker compose yaml
To allow the frontend team to develop against the backend without the neeed
for a real machine we are building a standalone backend running in a docker container
so that the websocket connection can be used with this dummy data.
As a first draft the container is running in host networking mode to allow
for easy connection to localhost.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>