Add config manager to replace most hardcoded values in the future
So far the backend is not persisting any data. Instead it is running completely
on hardcoded values.
This is no longer possible with the different WifiModes which are now no longer
treated on system level but treated by the backend. Therefore the backend needs
to be able to persist settings.
We are doing so by writing to a yaml file which we can load from disk again.
To be able to easily extend the config file we are also adding a default config
which in return is able to be merged with the read config from disk.
This allows us not only to add new settings flags without a lot of extra effort
but also to recover in case of a corrupted config file which will then contain
known-good default values. Further config sanity checks can be implemented in
the future.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>