config: Add Wifi Access point config
The current generation of meticulous software starts the machine into
Access Point mode.
Going forward we will need to be able to configure this behaviour dynamically.
To be able to do so we are adding the wifi hotspot configuration to the WifiManager
and create the hotspot according to the configuration.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>