api: Add Zeroconf announcement class
The meticulous machine is currently only reachable if the use
knows its IP Address. This is a mayor downside in many scenarios as not
all users are aware of their network setup and IPv6 addresses can look
suprisingly scary to the untrained eye.
To remidiate this issues we can use autodiscovery services, in this case zeroconfg.
The meticulous backend is providing a REST API via HTTP and a socket.io interface
on top of websockets.
As it is a standardized service type we are _not_ creating our own service type
like `_websocket._tcp.local` but instead use _http._tcp.local. .

The announced service can be seen in the network and be dumped by tools like avahi-browse
like such:

$ avahi-browse -l --all | grep met -i
+  wlan0 IPv4 MeticulousEspresso                            Web Site             local

Signed-off-by: Mimoja <mimoja@meticuloushome.com>