wifi: Check network manager running on startup
In some circumstances we can not call nmcli and it will fail instead.
This can either be cause by NetworkManager not being installed or it
being unavailable to connect to as in the case of a docker deployment.
We therefore error-check one nmcli command and use the result to abort on
all network operations.
As we can force the backend to be started after networking is online via its
systemd service we can be sure that we are not running into false positives
where waiting a short moment for NetworkManager to start would have unblocked
us. We are guaranteed that NM is running.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>