config: Add sensor logging setting to the config
So far the backend logs all sensor data during an action or when the operator
has explicitly requested it via "show". In headless deployments this flag cannot be
passed as no pty is connected to stdin.
To still be able to debug with all ESP32 messages we persist the setting in the config
and attach the "show" and "hide" commands to it so that they are presisted over
restarts including crashes.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>