api: Add Notification message logic
We want to be able to send notifications to the dial app and the different frontend.
To be able to do so we implement a dataclass containing a message.
Furthermore we add a NotificationManager which keeps track of unacknowledged
notifications and can look a few minutes into the past.
We also provide a REST API for reconnecting frontends to fetch all open
messages.
The logic can be tested via the newly added "notification" console command
which will send a single-button notification.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>