api: Add Profile manager logic
So far the dial app and the react dash app had their own profile storage.
This lead to desyncronization issues and the inablity to edit the profile
which was loaded on one of them on the other.
To remidiate this we are now moving all profile logic into the backend.
As a first start we add a super simple read/write/list endpoint which safes
the profile to disk.
Longterm this sounds like a great usecase of a database. But making the backend
stateless is currently far out so we build the MVP solution first to be able to
prototype and work from there.
It is noteably that the backend _does not_ understand the profiles so far and
is therefore unable to verify them at all.
It also doesn't sent the profile to the ESP32 yet once it has been writen

Signed-off-by: Mimoja <mimoja@meticuloushome.com>