api: Add update endpoint
The machines firmware needs to be updateable for development purposes. Furthermore
we want to be able to update the machine quickly in the field in case of errors without
having to update the full emmc image first.
Add an firmware update handler which unpacks the update package and flashes it to the ESP32.
FIXME: Flashing is not yet implemented.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>