wifi: Add HostnameManager class to update the hostname on startup
So far all variscite base boards are provisioned with the hostname `imx8mn-var-som`.
This is not only non-unique in any ways but also really confusing to customers.
Instead of provisioning every machine by hand we set the hostname in the backend and
reproducible.
We are using the last two bytes of the WIFI MAC address which should give us reasonably
good collision prevention in almost all scenarios. Collisions in itself are not the worst
per-se but they break zeroconf and can be anoying.
To be on the safe side we are announcing the machines IP on top of the hostname anyways.

To get to a name we are using a lookup table consisting of 255 adjectives and 255 nouns roughly
related to coffee making and the consumption thereof.
This gives a heroku alike unique name generator which generated the same hostname on a given
SOM every time.
We therefore never need to set the hostname on machine provisioning and can ship one uniform
image as an update without ever running risk of resetting the hostname on accident.

Example Hostnames are:
- meticulous-epicEspressoMartini
- meticulous-heroicCafeCrema
- meticulous-virtuosoCoffeePuck
- meticulous-idolizedFlavorProfile
- meticulous-energizingFlatWhite
- meticulous-artisanalNetwork
- meticulous-psychicMilk

Signed-off-by: Mimoja <mimoja@meticuloushome.com>