Fix: Revert machine status to string to allow 1:1 passthrough
In ```d695a6e sensors: Introduce data classes for sensor data parsing```
we introduced the MachineStateEnum for represent the machine state.
However we underestimated that the profile could basically set new status
messages to communicate with the user on the dial app.
To unblock this data workflow for now until we can come up with something
better we remove the MachineStateEnum again and use a const loopup instead

Signed-off-by: Mimoja <mimoja@meticuloushome.com>