wifi: Set nmcli LANG to C.UTF-8
Linux locales are a set of environment variables that define
the language and regional formatting for a user in a Linux system.
These variables determine how the system presents and interprets
data such as time, date, currency, and character encoding.

By default the nmcli modules enforces the LANG to be "C".
The Linux locale C is a special locale setting that represents
the default locale in Unix and Linux systems.
It is the least common denominator for software running on a Linux / Unix
system.
For the C locale the system uses ASCII as the character encoding
which breaks the decoding of WIFI SSIDs for us.
Set the locale to C.UTF-8 instead.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>