api: Host a minimal graph debug UI as well as then dashboard
The debug ui we have for internal testing is not easily adapted for use with the
backend, which is why we are setteling for just the graph viewer.
On top of that the dashboard is not going to be completely sunset and we will instead
host it on the machine as well.
While the watcher is currently hosting it on port 3000 for consistency we are now also hosting
it on the machine itself.
In the future we should evaluate putting a reverse proxy in place instead.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>