wifi/zeroconf: Allow manual overwrite of the zeroconf IP and hostname for testing
The Zeroconf service announcement is designed to announce the current
networking state of the machine. While announcing works inside a docker container
without additional config the container itself is behind the docker network and can
therefore only fetch limited informations about the outside world.
As we are running the container on multiple platforms we should not make to many
assumptions about the network layout and instead offer a simpe overwrite environment
variable which can be used to manipulate the announement from the machine instead.

Setting ZEROCONF_OVERWRITE=192.168.2.123/24,TestServer
will create the following announcement:

  Name: meticulous-spicyCafeAuLait._http._tcp.local.
  Addresses: 192.168.2.123:8080
  Weight: 0, priority: 0
  Server: TestServer.
  Properties are:
    b'server_name': b'TestServer'

The passing of the server name is necessary as some zeroconf implementations do
to support the full SRV message and try to resolve the server name in the passed
domain.
In the example above this means that e.g. avahi-browse will try to resolve
TestServer.local via mDNS instead of just using the passed IP address directly.
We therefore allow for overwriting of both.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>