ble: Always start the improv gatt server on backend start
The backend is supposed to allow the user to connect to different wifis.
This is particulary important as the user has no other interface to
create profiles at this point outside of the mobile app and eventually the
web app.
To be able to test the improv implementation and to not wait for the system
to propagate the new announcements we are now starting the GATT server on
backend start.
This comes with an added challenge as the variscite-bluetooth.service is
actively messing with the state of the bluetooth module after boot which leads
to the modules being reported ready just to be then powered off by firmware.
To prevent running into issues with this behaviour we are instead waiting for
at least 15 seconds of uptime before we interact with the bluetooth module.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>