Controllers Class Added
The Controllers class has been added to the file with the same name, and it now works independently. Single quotes ' have been changed to double quotes " in the Triggers file, with the aim of standardizing the dictionaries of Profile Manager.

The way triggers are printed has been modified to make them easier to read