shots: Add debug shot manager and history API
We often times want to see the full shot including heatup of and purge and everything.
We therefore add the ShotDebugManager which is almost a perfect copy of the ShotManager
but records to CSV instead and into a different folder.

To access all the shots we build a history conversion tool which automatically uncompresses
the files on transfer.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>