profile_converter: fix limit types
The limit types we were emitting were not in line with the complex JSON
spec. We therefore had to update them to make the ESP happy-ish.
It is worth noting that the first sending of the complex JSON still
fails to unknown reasons, we speculate it could be due to the serial
console.

Signed-off-by: Mimoja <mimoja@meticuloushome.com>