profiles: Allow for reuse of legacy italian profiles if they have the type field
We currently cannot handle the legacy profiles even tho the firmware
is - at least right now - still able to execute them.
We therefore allow the previous backend logic where the dial app
and dashboard can stread the old profile if and only if it is properly
marked as type:legacy.
The socket.io feedProfile logic was left untouched for this

Signed-off-by: Mimoja <mimoja@meticuloushome.com>