profiles: Preprocess profile variables
In order to be able to use profile variables we need to parse
and inject them in the json. We do so by parsing them into a map
and whenever we encounter a string in the profile we replace it with
the variables value. Due to the stringy nature of json we can enforce variables
to have types which correspond to the value they are replacing will allow
frontends to assign them min and max values.
On the flipside this stringy nature makes the processing of the json computationally
intense which is why we are doing it in the backend and not the firmware

In total the following types will have to be supported for variables by a frontend
implementing variable editing:

- "time"
- "weight"
- "pressure"
- "flow"
- "piston_position"
- "temperature"
- "power"

We are _not_ supporting putting the interpolation type into a variable (yet)
even tho we initially discussed doing so. This way we limit the variable values to
numbers only for the first implementation.

We are also adding a unit test for the preprocessor which can be run with
`python -m unittest test_profile_preprocessor.py`

Signed-off-by: Mimoja <mimoja@meticuloushome.com>