Fix API history endpoint
The redirect from `history` to `history/` and from `history/debug` to
`history/debug/` got broken when the transition to the versioned API,
the URL used to redirect was not taking into account the version prefix
as it was added to the URL on the "get_routes()" mothod of the API class
and the redirect was manually added to the request handler.

The previous location of the debug logs hid a bug using the `history/`
endpoint. When trying to access the `history/debug` endpoint, request
was catched by the `history` endpoint due to the `history/(.*)` regex
so the directory looked for was `meticulous-user/history/debug` which is
no longer existing.

This commit fixes the redirect by getting the lates API version prefix
and appendig it to the URL used to redirect, also adds an exclusion
regex for the `debug` pattern on the `history endpoint regex`, whith
this, the debug endpoint is now accesible at `history/debug`.

The `ZstdDebugHandler` class was removed, the `debug` endpoint is now
hanlded also by `ZstdHistoryHandler`