fix: start the dbus service before launching the backend service
The base image does not have the dbus nor the bluez packages installed
which are required for rauc notifications as well as bluez interface collection

both packages are installed in the Docker imaged build process.

A new entry point is defined to start the dbus service before launching the
backend service.

There is still an issue with the bluetooth, it cannot be activated so the Gatt
server crashes