fix: multi update settings via SettingsHandler
When usind the SettingsHandler, the workflow should allow to modify
multiple settings with one request, but rn if any of the provided
settings fail to update correctly, any further setting changes are
aborted as the function returns.

Now, the function can handle individual failure cases for the already
defined settings. All the errors encountered are reported back in the
request response body and the status code for the request can vary from
`200 (OK)` when no settings encounter errors, `207 (Multi-status)` when at
least 1 setting was successfully updated and `404 (not found)` when none
of the settings were successfully updated

In order to avoid the method USBManager.setUSBMode() to rely on the
loaded value of the MeticulousConfig dictionary -that made necessary to
update it first before calling it- the value is passed as a parameter to
the function, also in case that the value is not valid, a RuntimeError is
raised