fix: debug logs are not being saved after a shot
Within the Machine class, there is a flag used to see if the shot really has started, meaning the
valve has closed, this is used to only save the shot data when a shot has really started which works on the
ShotManager but not with the ShotDebugManager.

The reason might be that the flag is requred to be set to run ShotDebugManager.stop() in charge of saving data
and the machine state to be idle or purge. But the flag seems to be cleared before the machine enters any of those
states, while retracting and finding a stable weight, this prevents the condition to be met and the stop method is not
called.

Now, we keep track of the shot status inside the ShotDebugManager itself, by setting a flag when the shot has really started
we dont need the external flag that can be cleared by other means anymore so we only focus on the machine state to call the
stop() method. And if we have not begun the shot, we discard the collected data, else we proceed to save it on disk