feat: save logs happening within a shot timeframe into the debug shot file
A ShotLogHandler handler was created to also send the logging messages to an array of log messages in
the ShotDebugManager class. This handler is added to all existing loggers when a debug shot is started
and removed once the debug shot is stopped. While the handler is active in the loggers, they will append
an object like the following, for every message processed, into the mentioned array.

``` json
{
    "profile_ms": int,
    "caller": str,
    "log_message": str,
}
```

where
`profile_ms` is the time respect the start of the profile in ms that the log was created.
`caller` is the name of the logger instance that created the log e.g(`machine` or `tornado.access`).
`log_message` is the message logged

Finally that list of dicts is copied to the DebugShot object to be saved in the json debug file, and then cleared