refactor: move add/remove log handles to log.py add handleLog to ShotDebugManager
functions to add/remove a handler to the loggers now form part of MeticulousLogger,
they also allow pass the name of a logger to only add/remove the handler to it, if no
name is passed it adds/removes it to all available loggers.

ShotLogHandler.emit() now calls the new function SHotDebugManager.handleLog(). With this
we dont need to remove the handler as the handleLog fnc takes care of only saving logs when
there is a deebug shot.